Rename video list activity and put another hierarchy for video activities
