+++ app/src/main/res/layout/activity_video_display.xml
@@ -8,7 +8,7 @@
     android:paddingRight="16dp"
     android:paddingTop="16dp"
     android:paddingBottom="16dp"
-    tools:context="com.team03.coursewarehub.VideoDisplay">
+    tools:context="com.team03.coursewarehub.Youtube.VideoDisplay">
 
     <ScrollView
         android:layout_width="fill_parent"
@@ -31,6 +31,16 @@
             android:layout_alignParentTop="true"
             android:layout_gravity="center"
             android:layout_marginTop="10dp"/>
+
+
+        <TextView
+            android:layout_width="wrap_content"
+            android:layout_height="wrap_content"
+            android:id="@+id/textView"
+            android:fontFamily="sans-serif-condensed"
+            android:textStyle="italic"
+            android:textSize="40sp"
+            android:layout_gravity="center_horizontal" />
 
         <com.google.android.youtube.player.YouTubePlayerView
             android:id="@+id/youtube_view"
