+++ app/src/main/res/menu/menu_search_second_you_tube.xml
@@ -1,6 +1,6 @@
 <menu xmlns:android="http://schemas.android.com/apk/res/android"
     xmlns:tools="http://schemas.android.com/tools"
-    tools:context="com.team03.coursewarehub.SearchSecondVideoActivity">
+    tools:context="com.team03.coursewarehub.Youtube.VideoListActivity">
     <item android:id="@+id/action_settings" android:title="@string/action_settings"
         android:orderInCategory="100" android:showAsAction="never" />
 </menu>
